Update page.tsx
I've made the following changes to the quest dashboard page:

Changed the rewardButtonTitle from a conditional "Enable"/"Disable" to a static "Analytics"

Updated the onRewardButtonClick handler to:

Remove the quest enable/disable functionality

Add new functionality to open the analytics page in a new tab using the quest ID

Use window.open() with _blank target to ensure it opens in a new tab

The changes are now complete and match the requirements.

The "Enable/Disable" button has been replaced with an "Analytics" button that will open /analytics/${questData.id} in a new tab when clicked.